Tests added: rsx/rsx_fp_dynamic
I don't know if these tests will be useful at all. They are just the
same as the rsx_fp_static tests, but the fragment shader receives
parameters such as the elapsed time (float iGlobalTime) and a pointer
controlled by the gamepad (float2 iGamePad). That allows the shader to
draw different things on each frame so images do not make a lot of sense
when it comes to comparing results. So, at the moment this folder is
nothing but just eye candy.